tried to add XCTest to project
